Separated AttackConfig from NPCConfigs.
