+++ source/core/src/test/com/csse3200/game/components/npc/attack/AOEAttackComponentTest.java
@@ -4,7 +4,7 @@ import com.csse3200.game.areas.GameAreaService;
 import com.csse3200.game.areas.MainGameArea;
 import com.csse3200.game.components.CombatStatsComponent;
 import com.csse3200.game.entities.Entity;
-import com.csse3200.game.entities.configs.NPCConfigs;
+import com.csse3200.game.entities.configs.AttackConfig;
 import com.csse3200.game.extensions.GameExtension;
 import com.csse3200.game.services.GameTime;
 import com.csse3200.game.services.ServiceLocator;
@@ -116,9 +116,9 @@ class AOEAttackComponentTest {
 
     private Entity createAttacker(Entity target, float aoeRadius) {
         // Setup attacker configs
-        NPCConfigs.NPCConfig.EffectConfig[] effectConfigs = {}; // No effects
-        NPCConfigs.NPCConfig.AttackConfig.AOEAttack aoeAttackConfig
-                = new NPCConfigs.NPCConfig.AttackConfig.AOEAttack();
+        AttackConfig.EffectConfig[] effectConfigs = {}; // No effects
+        AttackConfig.AOEAttack aoeAttackConfig
+                = new AttackConfig.AOEAttack();
         aoeAttackConfig.range = aoeRadius;
         aoeAttackConfig.rate = 1f;
         aoeAttackConfig.effects = effectConfigs;
