+++ source/core/src/test/com/csse3200/game/components/npc/attack/AttackComponentTest.java
@@ -4,7 +4,7 @@ import com.csse3200.game.areas.GameAreaService;
 import com.csse3200.game.areas.MainGameArea;
 import com.csse3200.game.components.CombatStatsComponent;
 import com.csse3200.game.entities.Entity;
-import com.csse3200.game.entities.configs.NPCConfigs;
+import com.csse3200.game.entities.configs.AttackConfig;
 import com.csse3200.game.extensions.GameExtension;
 import com.csse3200.game.services.GameTime;
 import com.csse3200.game.services.ServiceLocator;
@@ -92,7 +92,7 @@ class AttackComponentTest {
     }
 
     private Entity createAttacker(Entity target) {
-        NPCConfigs.NPCConfig.EffectConfig[] effectConfigs = {}; // No effects for simplicity
+        AttackConfig.EffectConfig[] effectConfigs = {}; // No effects for simplicity
         Entity attacker = new Entity()
                 .addComponent(new CombatStatsComponent(10, 10))
                 .addComponent(new AttackComponent(target, 2f, 1f, effectConfigs) {
