+++ source/core/src/test/com/csse3200/game/components/npc/attack/MeleeAttackComponentTest.java
@@ -4,7 +4,7 @@ import com.csse3200.game.areas.GameAreaService;
 import com.csse3200.game.areas.MainGameArea;
 import com.csse3200.game.components.CombatStatsComponent;
 import com.csse3200.game.entities.Entity;
-import com.csse3200.game.entities.configs.NPCConfigs;
+import com.csse3200.game.entities.configs.AttackConfig;
 import com.csse3200.game.extensions.GameExtension;
 import com.csse3200.game.services.GameTime;
 import com.csse3200.game.services.ServiceLocator;
@@ -98,9 +98,8 @@ class MeleeAttackComponentTest {
 
     private Entity createAttacker(Entity target) {
         // Setup attacker configs
-        NPCConfigs.NPCConfig.EffectConfig[] effectConfigs = {}; // No effects
-        NPCConfigs.NPCConfig.AttackConfig.MeleeAttack meleeAttackConfig
-                = new NPCConfigs.NPCConfig.AttackConfig.MeleeAttack();
+        AttackConfig.EffectConfig[] effectConfigs = {}; // No effects
+        AttackConfig.MeleeAttack meleeAttackConfig = new AttackConfig.MeleeAttack();
         meleeAttackConfig.range = 2f;
         meleeAttackConfig.rate = 1f;
         meleeAttackConfig.effects = effectConfigs;
